pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-converter"
version = "1.0.0"
description = "Convert PNG and JPEG images to AVIF format"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["image", "converter", "avif", "png", "jpeg", "compression"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
    "Pillow>=10.0.0",
    "pillow-avif-plugin>=1.4.0",
]

[project.scripts]
image-converter = "image_converter.cli.app:main"
img2avif = "image_converter.cli.app:main"

[project.urls]
Homepage = "https://github.com/yourusername/image-converter"
Repository = "https://github.com/yourusername/image-converter"

[tool.setuptools.packages.find]
where = ["src"]

